Gem sorting is important
This notes that the `responders` gem must be before the `roar-rails` gem in the gemfile. I don't know if this is the case for Rails < 5, but it's the case running the current Rails 5 alpha. I don't see how it can hurt in earlier versions of Rails, so I think the change is pretty safe.