Call model.model_name.name instead of model.class.name to determine responder path.
ActiveModel supports a #model_name method which returns a class providing information
about the model name, and removes the need to infer this from the class heirarchy.

The default model_name can also be over-ridden, which is especially useful to have
rails treat all sub-classes of an STI base model the same (using the base path) for
view and other related file lookups.

See also http://stackoverflow.com/questions/29527156/rails-sti-override-model-name-in-parent-class-for-all-subclasses